service/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
boto3 = "^1.26.109"
openai = {extras = ["embeddings", "wandb", "datalib"], version = "^0.27.4"}
playwright = "^1.32.1"
python = "^3.10"
boilerpy3 = "^1.0.6"
beautifulsoup4 = "^4.12.2"
retrying = "^1.3.4"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
uvicorn = "^0.21.1"
websockets = "^11.0.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^23.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
